dummy button of new memo
todo:
Before adding the ability to press a button to create a new memo, there is one thing that needs to be done.
That is to modify the memo itself so that it has a structure that allows it to be structured and saved.
At the moment, Messages are structured, but Memos (temporary) are still completely dummy.
Let's finish this first.